Basic Socket Actions Passing API Data to Clients
Creates the ability for clients to add new stocks and remove stocks
which are already selected through the application. State is maintained
in memory on the server. Currently stocks to be added and removed are
hard coded. This will be changed in a future commit once the frontend
has been developed further.