Error Handling Incorrect Stocks and Adding Dups
Implements error handling in the API so as to return an error if the
response from the server contains an error message. Additionally adds
a new function inside the socket actions to check if the stock is
already in the data array. Removes the initial check to see if the
data array is empty or not and will only return the data array when
specific stocks are requested. Prevents from adding duplicate stocks
to the data array. Adds a new socket listener to recieve an error if
a bad stock code is chosen.