Adds Color Support to the Line Chart
Adds color support to lines graphed on the chart. Twelve lines can be
graphed simultaneously before colors are reused. A constant color array
was added to the store so that the chart element and the stock list
element could access the same colors. Colors and stocks are related by
index in which they appear in the stocks array and colors array.