Adds New DataPoints Component
Adds a new DataPoints component to show each data point on the line.
Most of the display attributes are customizable through props. Lines
and datapoints are grouped into the SVG under a common g element.
Will be utilized to add mouseOver events so that each datapoint
can be individually inspected.