Adds ToolTip Component to Chart
Adds a tooltip component so that users can view information about
individual datapoints. Mouseover events are added to each datapoint and
the chart contains the actions for mouseover / mouseout. The tooltip's
position and visibility is toggled through a change in state of the
chart component. The tooltip finally renders itself based off props.